requirements.txt
Django==5.2.7
djangorestframework==3.20.0
djangorestframework-simplejwt==6.0.0
drf-yasg==1.22.6
PyJWT==2.8.0
